Added blog entry and updated page headings
Added a blog entry about the path to ruby mastery.

Also, I noticed that I was using <span> tags for all of the page titles when
they should really be <h1> elements so that they will get higher page rankings.

On that same note, I added some meta tags to describe the: author,
description, and keywords