Added ability to places notes on any page
This will be useful for blog pages or random-entry pages that have an
update later on or just a disclaimer at the top of the page. All that
is required is to match the name of hte log entry or the random-entry
file and append it with '.notes'. This is a text file and is read-in
by the sinatra-app.rb file when the page is request (if the notes are
available).

Right now they are plain text files, however I'd like to upgrade them
sometime in the future to being markdown files for a little more
expression.