Replaced custom code-highlighting with Gists
In a move to clean up my web-application, I'm moving away from doing any
type of code-formatting myself. It's just ugly code and more headache than
I'm willing to deal with when working on my personal site.

Gists are pretty (in format) and are automatically versioned for me. There
is also a commenting facility if anyone is so inclined (I'm for sure not
going to put that functionality on my site).

+ Updated Fnord Metrics blog
+ Updated WCF API Versioning blog