General Cleanup/Refactor of Site
My site was an ugly mess that I put together when I
first learned what Sinatra was. Looking back at it now,
it's kind of embarrasing. No I didn't add anything new
to the site or really improve it's functionality in any
way whatsoever. However, the code is cleaner and that
makes me happier! And it means I can actually tell
people to check out the source for my website without
having to give any (or at least less) disclaimers.

The refactoring that was done includes:
  + Re-organizing the Gemfile (into proper groups)
  + updating the main app-file to not use 'classic' mode
  + Removed useless config files and moved all config
    options into the new 'non-classic?' app.
  + altered loading of route-files to be more modular with
    new app layout
  + Removed the 'prettyfier' helper as I'm now using Gists
    for all of my snippets (not needed anymore)
  + Added newrelic.yml config for heroku (for better logging)
  + Renamed app-file to be more common ('app.rb')
  + Updated require's to use bundler-loading based on
    current rack-env setting.